+++ packages/_private/vue-demo/src/AllComponentParts/DemoMenu.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { SynVueMenu, SynVueMenuItem, SynVueDivider } from '@synergy-design-system/vue';
+import { SynVueInput, SynVueMenu, SynVueMenuItem, SynVueDivider } from '@synergy-design-system/vue';
 </script>
 <template>
   <div style="width: 200px">
@@ -11,6 +11,31 @@ import { SynVueMenu, SynVueMenuItem, SynVueDivider } from '@synergy-design-syste
       <SynVueMenuItem value="copy">Copy</SynVueMenuItem>
       <SynVueMenuItem value="paste">Paste</SynVueMenuItem>
       <SynVueMenuItem value="delete">Delete</SynVueMenuItem>
+    </SynVueMenu>
+  </div>
+
+  <SynVueDivider />
+
+  <div
+    data-testid="menu-1295-steals-focus"
+    style="
+      display: flex;
+      flex-direction: column;
+      gap: var(--syn-spacing-small);
+      margin: var(--syn-spacing-large) 0;   
+    "
+  >
+    <SynVueInput
+      name="search"
+      label="#1295: Focus should not be lost when hovering over the menu"
+      type="text"
+    />
+    <SynVueMenu>
+      <SynVueMenuItem value="undo">Undo</SynVueMenuItem>
+      <SynVueMenuItem value="redo">Redo</SynVueMenuItem>
+      <SynVueDivider />
+      <SynVueMenuItem value="cut">Cut</SynVueMenuItem>
+      <SynVueMenuItem value="copy">Copy</SynVueMenuItem>
     </SynVueMenu>
   </div>
 </template>