+++ packages/metadata/data/layers/full/component/component:syn-optgroup/components/optgroup.component.ts
@@ -120,6 +120,8 @@ export default class SynOptgroup extends SynergyElement {
           if (optgroupMutation.type === 'childList') {
             optgroupMutation.addedNodes.forEach((node) => {
               if (node instanceof HTMLElement && node.matches('syn-option')) {
+                // False positive for instanceof check, but we know this is correct since the mutation observer is only looking for syn-options
+                // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
                 handleInitialDisabledForOption(node as SynOption, this.disabled);
               }
             });